Fixed logic based on Automata theory.
Search was working wrong, for example when you run code`
Input search pattern: synopsys
Input text to search: synoooopsys
The result was 10 but it is wrong result, the result will be "Pattern not found".

You could fix it simply comparing pattern`s and result length, but this would not fix the wrong operation of the theory Automata.
I just added some minor changes to the ComputeNextState function.